Moving classes in v2 packages to their parent package. There was a minor warning fix in BulkMutation as well.
