+++ bigtable-client-core/src/main/java/com/google/cloud/bigtable/grpc/scanner/AbstractBigtableResultScanner.java
@@ -13,10 +13,9 @@
  * See the License for the specific language governing permissions and
  * limitations under the License.
  */
-package com.google.cloud.bigtable.grpc.scanner.v2;
+package com.google.cloud.bigtable.grpc.scanner;
 
 import com.google.bigtable.v2.Row;
-import com.google.cloud.bigtable.grpc.scanner.ResultScanner;
 
 import java.io.IOException;
 import java.util.ArrayList;
