+++ bigtable-client-core/src/test/java/com/google/cloud/bigtable/grpc/scanner/StreamingBigtableResultScannerTest.java
@@ -29,7 +29,6 @@ import org.mockito.MockitoAnnotations;
 
 import com.google.bigtable.v2.Row;
 import com.google.cloud.bigtable.grpc.io.CancellationToken;
-import com.google.cloud.bigtable.grpc.scanner.v2.StreamingBigtableResultScanner;
 
 @RunWith(JUnit4.class)
 public class StreamingBigtableResultScannerTest {
